requirements.txt
timm
transformers
fairscale
pycocoevalcap
torch
torchvision
Pillow
